Added historical prediction logic

+++ src/agents/models/ApplianceConsumption.java
@@ -3,11 +3,11 @@ package agents.models;
 import java.time.DayOfWeek;
 
 public class ApplianceConsumption {
-    public float consuming;
+    public double consuming;
     public int hourOfDay;
     public DayOfWeek dayOfWeek;
 
-    public ApplianceConsumption(float consuming, int hourOfDay, DayOfWeek dayOfWeek) {
+    public ApplianceConsumption(double consuming, int hourOfDay, DayOfWeek dayOfWeek) {
         this.consuming = consuming;
         this.hourOfDay = hourOfDay;
         this.dayOfWeek = dayOfWeek;
